[level 0] Title: 자릿수 더하기, Time: 0.37 ms, Memory: 31.5 MB -BaekjoonHub

+++ 프로그래머스/lv0/120906. 자릿수 더하기/자릿수 더하기.cs
@@ -6,7 +6,7 @@ public class Solution {
         string tmp = n.ToString();
         for(int i = 0; i < tmp.Length; i++)
         {
-            answer += (int)tmp[i] - '0';
+            answer += tmp[i] - '0';
         }
         return answer;
     }
